# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_user.UpdateAvatarResponse:
    properties:
      avatar_url:
        type: string
      message:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Неверный формат запроса
        type: string
    type: object
...
  review.CreateReviewRequest:
    properties:
      author_name:
        example: Алексей
        type: string
      comment:
        example: Очень вкусно, доставили быстро!
        type: string
      rating:
        example: 5
        type: integer
    type: object
  review.Review:
    properties:
      author_name:
        example: Алексей
        type: string
      comment:
        example: Очень вкусно, доставили быстро!
        type: string
      created_at:
        example: "2026-04-20"
        type: string
      id:
        example: 1
        type: integer
      rating:
        example: 5
        type: integer
      restaurant_id:
        example: 1
        type: integer
    type: object
  review.ReviewsResponse:
    properties:
      reviews:
        items:
          $ref: '#/definitions/review.Review'
        type: array
      total:
        example: 2
        type: integer
    type: object
...
  user.UpdateRoleRequest:
    properties:
      new_role:
        type: string
      user_id:
        type: integer
    type: object
  user.UserProfileResponse:
    properties:
      avatar_url:
        example: users/avatars/fjaun99f.png
        type: string
      email:
        example: terminator2007@gmail.com
        type: string
      name:
        example: Илья
        type: string
    type: object
  user.UserProfileUpdateRequest:
    properties:
      email:
        example: new_mail@gmail.com
        type: string
      name:
        example: Андрей
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API бэкенда для проекта "Delivery Club" от команды NaNcats
  title: NaNcats Delivery API
  version: "1.0"
paths:
...
  /admin/users/role:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ключ
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.UpdateRoleRequest'
      produces:
      - application/json
      responses: {}
      summary: Смена роли пользователя (админ)
      tags:
      - admin
...
  /profile:
    get:
      consumes:
      - application/json
      description: Возвращает данные профиля (имя и email) текущего авторизованного
        пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение данных профиля
          schema:
            $ref: '#/definitions/user.UserProfileResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение профиля пользователя
      tags:
      - profile
    patch:
      consumes:
      - application/json
      description: Частично обновляет данные профиля текущего пользователя (имя и/или
        email)
      parameters:
      - description: Данные для обновления профиля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.UserProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль успешно обновлен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации JSON или нет данных для обновления
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Указанный email уже используется другим пользователем
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Обновление данных профиля
      tags:
      - profile
...
  /profile/avatar:
    delete:
      description: Удаляет аватар текущего авторизованного пользователя и устанавливает
        аватар по умолчанию.
      produces:
      - application/json
      responses:
        "200":
          description: Аватар успешно удален
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_user.UpdateAvatarResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Удаление аватара пользователя
      tags:
      - profile
    post:
      consumes:
      - multipart/form-data
      description: Загружает и обновляет аватар текущего авторизованного пользователя.
        Принимает multipart/form-data with полем 'avatar'.
      parameters:
      - description: Файл аватара (WEBP/JPG/JPEG/PNG, до 5МБ)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Аватар успешно обновлен
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_user.UpdateAvatarResponse'
        "400":
          description: Ошибка запроса (файл слишком большой, неверный формат или отсутствует)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Обновление аватара пользователя
      tags:
      - profile
...
  /restaurants/{id}/reviews:
    get:
      description: Возвращает список всех отзывов для конкретного ресторана по его
        ID
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение отзывов
          schema:
            $ref: '#/definitions/review.ReviewsResponse'
        "400":
          description: Неверный ID ресторана
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение списка отзывов
      tags:
      - review
    post:
      consumes:
      - application/json
      description: Создает новый отзыв для указанного ресторана с оценкой и комментарием
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для создания отзыва
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/review.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Отзыв успешно создан
          schema:
            $ref: '#/definitions/review.Review'
        "400":
          description: Неверный формат JSON, ID ресторана или ошибка валидации полей
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Создание отзыва
      tags:
      - review
...
swagger: "2.0"